Do not use user config file when it doesn't exists (#737)
It caused startup fail. The core reason is that /rw is local to the VM,
so when package installed in TemplateVM, files placed to /rw aren't
visible to the TorVM (based on that template).
So check if the file exists before passing it to tor arguments and
completely remove it from the package.

BTW move its expected location to /rw/config.